update
A system that connects to more than one database through a form from which you dictate the
database data. Through this system, you can make the changes you want, and the system takes a
backup copy of the database that you made the change to before the change. 1/ You can make
inputs to help you with the changes you want to make. 2/ There is also a list of all activities that
occurred in the system. 3/ There are role and permissions for users to perform certain
operations on the system. 4/ Every change made to a database must be approved before it is
made, and the change is implemented on the database as soon as it is approved.4/ A backup copy
is taken of the database on which the change was made. 5/ A backup copy of the database that
was changed is taken before the change